MIR: Improves hip/cuda build times/memory by splitting out tests by policy
